requirements.txt
python-telegram-bot[job-queue]
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
pandas
fastapi
uvicorn
cryptography
python-multipart
apscheduler
